refactor(plugin): split response and its tests into per-domain files
- Split response.go into response_logs.go, response_stats.go, response_hits.go
- Split response_test.go likewise (response_logs/stats/hits_test.go)
- Keep shared types, error parsers and labelsToJSON in response.go
