+++ pkg/plugin/response.go
@@ -1,21 +1,11 @@
 package plugin
 
 import (
-	"bufio"
-	"bytes"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"hash/fnv"
 	"io"
-	"strconv"
-	"time"
 
-	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	"github.com/grafana/grafana-plugin-sdk-go/data"
-	"github.com/valyala/fastjson"
-
-	"github.com/VictoriaMetrics/victorialogs-datasource/pkg/utils"
 )
 
 const (
@@ -34,313 +24,20 @@ const (
 	logsVisualisation = "logs"
 )
 
-// buildLogID returns a stable identifier for a log row, used by Grafana's
-// Logs panel to enable per-row permalinks ("Copy shortlink").
-func buildLogID(ts time.Time, msg, stream string) string {
-	h := fnv.New64a()
-	_, _ = h.Write([]byte(strconv.FormatInt(ts.UnixNano(), 10)))
-	_, _ = h.Write([]byte{0})
-	_, _ = h.Write([]byte(msg))
-	_, _ = h.Write([]byte{0})
-	_, _ = h.Write([]byte(stream))
-	return strconv.FormatUint(h.Sum64(), 16)
+// Data contains fields
+// ResultType which defines type of the query response
+// Result resents json of the query response
+type Data struct {
+	ResultType string          `json:"resultType"`
+	Result     json.RawMessage `json:"result"`
 }
 
-var nowFunc = time.Now
-
-// parseStreamResponse reads data from the reader and collects
-// fields and frame with necessary information
-func parseInstantResponse(reader io.Reader) backend.DataResponse {
-
-	labelsField := data.NewFieldFromFieldType(data.FieldTypeJSON, 0)
-	labelsField.Name = gLabelsField
-
-	timeFd := data.NewFieldFromFieldType(data.FieldTypeTime, 0)
-	timeFd.Name = gTimeField
-
-	lineField := data.NewFieldFromFieldType(data.FieldTypeString, 0)
-	lineField.Name = gLineField
-
-	idField := data.NewFieldFromFieldType(data.FieldTypeString, 0)
-	idField.Name = gIDField
-
-	br := bufio.NewReaderSize(reader, 64*1024)
-	var parser fastjson.Parser
-	var finishedReading bool
-	for n := 0; !finishedReading; n++ {
-		b, err := br.ReadBytes('\n')
-		if err != nil {
-			if errors.Is(err, bufio.ErrBufferFull) {
-				backend.Logger.Debug("skipping line number: line too long", "lineNumber", n)
-				continue
-			}
-			if errors.Is(err, io.EOF) {
-				// b can be != nil when EOF is returned, so we need to process it
-				finishedReading = true
-			} else {
-				return newResponseError(fmt.Errorf("cannot read line in response: %s", err), backend.StatusInternal)
-			}
-		}
-
-		if len(b) == 0 {
-			continue
-		}
-
-		b = bytes.Trim(b, "\n")
-		value, err := parser.ParseBytes(b)
-		if err != nil {
-			return newResponseError(fmt.Errorf("error decode response: %s", err), backend.StatusInternal)
-		}
-
-		var (
-			rowMsg    string
-			rowStream string
-			rowTime   time.Time
-		)
-
-		if value.Exists(messageField) {
-			message := value.GetStringBytes(messageField)
-			rowMsg = string(message)
-			lineField.Append(rowMsg)
-		}
-		if value.Exists(timeField) {
-			t := value.GetStringBytes(timeField)
-			getTime, err := utils.GetTime(string(t))
-			if err != nil {
-				return newResponseError(fmt.Errorf("error parse time from _time field: %s", err), backend.StatusInternal)
-			}
-			rowTime = getTime
-			timeFd.Append(rowTime)
-		}
-
-		labels := data.Labels{}
-		if value.Exists(streamField) {
-			stream := value.GetStringBytes(streamField)
-			rowStream = string(stream)
-			stf, err := utils.ParseStreamFields(rowStream)
-			if err != nil {
-				return newResponseError(fmt.Errorf("%s", err), backend.StatusInternal)
-			}
-			for _, field := range stf {
-				labels[field.Label] = field.Value
-			}
-			if !value.Exists(messageField) && len(stf) >= 0 {
-				lineField.Append("")
-			}
-		}
-
-		obj, err := value.Object()
-		if err != nil {
-			return newResponseError(fmt.Errorf("error get object from decoded response: %s", err), backend.StatusInternal)
-		}
-		obj.Visit(func(key []byte, v *fastjson.Value) {
-			if bytes.Equal(key, []byte(timeField)) ||
-				bytes.Equal(key, []byte(streamField)) ||
-				bytes.Equal(key, []byte(messageField)) {
-				return
-			}
-			fieldName := string(key)
-			value := string(v.GetStringBytes())
-			labels[fieldName] = value
-		})
-
-		d, err := labelsToJSON(labels)
-		if err != nil {
-			return newResponseError(err, backend.StatusInternal)
-		}
-		labelsField.Append(d)
-		idField.Append(buildLogID(rowTime, rowMsg, rowStream))
-	}
-
-	// Grafana expects lineFields to be always non-empty.
-	if lineField.Len() == 0 {
-		for i := 0; i < labelsField.Len(); i++ {
-			label := labelsField.At(i)
-			lineField.Append(fmt.Sprintf("%s", label))
-		}
-	}
-
-	// Grafana expects time field to be always non-empty.
-	if timeFd.Len() == 0 {
-		now := nowFunc()
-		for i := 0; i < lineField.Len(); i++ {
-			timeFd.Append(now)
-		}
-	}
-
-	frame := data.NewFrame("", timeFd, lineField, idField, labelsField)
-
-	rsp := backend.DataResponse{}
-	frame.Meta = &data.FrameMeta{}
-	rsp.Frames = append(rsp.Frames, frame)
-
-	return rsp
-}
-
-// parseStreamResponse reads data from the reader and collects
-// fields and frame with necessary information
-// it looks like the parseInstantResponse function, but it reads data and continuously
-// parse the lines from the reader and we need to collect only one data.Frame
-func parseStreamResponse(reader io.Reader, ch chan *data.Frame) error {
-
-	br := bufio.NewReaderSize(reader, 64*1024)
-	var parser fastjson.Parser
-	var finishedReading bool
-	for n := 0; !finishedReading; n++ {
-		labelsField := data.NewFieldFromFieldType(data.FieldTypeJSON, 0)
-		labelsField.Name = gLabelsField
-
-		timeFd := data.NewFieldFromFieldType(data.FieldTypeTime, 0)
-		timeFd.Name = gTimeField
-
-		lineField := data.NewFieldFromFieldType(data.FieldTypeString, 0)
-		lineField.Name = gLineField
-
-		idField := data.NewFieldFromFieldType(data.FieldTypeString, 0)
-		idField.Name = gIDField
-
-		b, err := br.ReadBytes('\n')
-		if err != nil {
-			if errors.Is(err, bufio.ErrBufferFull) {
-				backend.Logger.Debug("skipping line number: line too long", "lineNumber", n)
-				continue
-			}
-			if errors.Is(err, io.EOF) {
-				// b can be != nil when EOF is returned, so we need to process it
-				finishedReading = true
-			} else {
-				return fmt.Errorf("cannot read line in response: %s", err)
-			}
-		}
-
-		if len(b) == 0 {
-			continue
-		}
-
-		b = bytes.Trim(b, "\n")
-		value, err := parser.ParseBytes(b)
-		if err != nil {
-			return fmt.Errorf("error decode response: %s", err)
-		}
-
-		var (
-			rowMsg    string
-			rowStream string
-			rowTime   time.Time
-		)
-
-		if value.Exists(messageField) {
-			message := value.GetStringBytes(messageField)
-			rowMsg = string(message)
-			lineField.Append(rowMsg)
-		}
-		if value.Exists(timeField) {
-			t := value.GetStringBytes(timeField)
-			getTime, err := utils.GetTime(string(t))
-			if err != nil {
-				return fmt.Errorf("error parse time from _time field: %s", err)
-			}
-			rowTime = getTime
-			timeFd.Append(rowTime)
-		}
-
-		labels := data.Labels{}
-		if value.Exists(streamField) {
-			stream := value.GetStringBytes(streamField)
-			rowStream = string(stream)
-			stf, err := utils.ParseStreamFields(rowStream)
-			if err != nil {
-				return err
-			}
-			for _, field := range stf {
-				labels[field.Label] = field.Value
-			}
-		}
-
-		obj, err := value.Object()
-		if err != nil {
-			return fmt.Errorf("error get object from decoded response: %s", err)
-		}
-		obj.Visit(func(key []byte, v *fastjson.Value) {
-			if bytes.Equal(key, []byte(timeField)) ||
-				bytes.Equal(key, []byte(streamField)) ||
-				bytes.Equal(key, []byte(messageField)) {
-				return
-			}
-			fieldName := string(key)
-			value := string(v.GetStringBytes())
-			labels[fieldName] = value
-		})
-
-		d, err := labelsToJSON(labels)
-		if err != nil {
-			return err
-		}
-		labelsField.Append(d)
-		idField.Append(buildLogID(rowTime, rowMsg, rowStream))
-
-		// Grafana expects lineFields to be always non-empty.
-		if lineField.Len() == 0 {
-			for i := 0; i < labelsField.Len(); i++ {
-				label := labelsField.At(i)
-				lineField.Append(fmt.Sprintf("%s", label))
-			}
-		}
-
-		// Grafana expects time field to be always non-empty.
-		if timeFd.Len() == 0 {
-			now := nowFunc()
-			for i := 0; i < lineField.Len(); i++ {
-				timeFd.Append(now)
-			}
-		}
-
-		frame := data.NewFrame("", timeFd, lineField, idField, labelsField)
-		// this is necessary information because the logs visualization is preferred
-		frame.Meta = &data.FrameMeta{PreferredVisualization: logsVisualisation}
-
-		ch <- frame
-	}
-
-	return nil
-}
-
-func parseStatsResponse(reader io.Reader, q *Query) backend.DataResponse {
-	var rs Response
-	if err := json.NewDecoder(reader).Decode(&rs); err != nil {
-		err = fmt.Errorf("failed to decode body response: %w", err)
-		return newResponseError(err, backend.StatusInternal)
-	}
-	rs.ForAlerting = q.ForAlerting
-
-	frames, err := rs.getDataFrames()
-	if err != nil {
-		err = fmt.Errorf("failed to prepare data from response: %w", err)
-		return newResponseError(err, backend.StatusInternal)
-	}
-
-	for i := range frames {
-		q.addMetadataToMultiFrame(frames[i])
-		q.addIntervalToFrame(frames[i])
-	}
-
-	return backend.DataResponse{Frames: frames}
-}
-
-func parseHitsResponse(reader io.Reader) backend.DataResponse {
-	var hr HitsResponse
-	if err := json.NewDecoder(reader).Decode(&hr); err != nil {
-		err = fmt.Errorf("failed to decode body response: %w", err)
-		return newResponseError(err, backend.StatusInternal)
-	}
-
-	frames, err := hr.getDataFrames()
-	if err != nil {
-		err = fmt.Errorf("failed to prepare data from response: %w", err)
-		return newResponseError(err, backend.StatusInternal)
-	}
-
-	return backend.DataResponse{Frames: frames}
+// Response contains fields from query response
+type Response struct {
+	Status      string `json:"status"`
+	Data        Data   `json:"data"`
+	Error       string `json:"error"`
+	ForAlerting bool   `json:"-"`
 }
 
 // parseErrorResponse reads data from the reader and returns error
@@ -383,223 +80,4 @@ func labelsToJSON(labels data.Labels) (json.RawMessage, error) {
 	}
 
 	return b, nil
-}
-
-const (
-	vector, matrix = "vector", "matrix"
-)
-
-// Result represents timeseries from query
-type Result struct {
-	Labels Labels  `json:"metric"`
-	Values []Value `json:"values"`
-	Value  Value   `json:"value"`
-}
-
-// Value represents timestamp and value of the timeseries
-type Value [2]interface{}
-
-// Labels represents timeseries labels
-type Labels map[string]string
-
-// Data contains fields
-// ResultType which defines type of the query response
-// Result resents json of the query response
-type Data struct {
-	ResultType string          `json:"resultType"`
-	Result     json.RawMessage `json:"result"`
-}
-
-// Response contains fields from query response
-type Response struct {
-	Status      string `json:"status"`
-	Data        Data   `json:"data"`
-	Error       string `json:"error"`
-	ForAlerting bool   `json:"-"`
-}
-
-// logStats represents response result from the
-// stats endpoints of the VictoriaLogs
-type logStats struct {
-	Result []Result `json:"result"`
-}
-
-func (ls logStats) vectorDataFrames() (data.Frames, error) {
-	frames := make(data.Frames, len(ls.Result))
-	for i, res := range ls.Result {
-		value := res.Value
-
-		ts, err := getTimestamp(value[0])
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse timestamp for metric %v: %w", res, err)
-		}
-
-		valuePtr, err := getFloatPtr(value[1])
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse float value for metric %v: %w", res, err)
-		}
-
-		frames[i] = data.NewFrame("",
-			data.NewField(data.TimeSeriesTimeFieldName, nil, []time.Time{ts}),
-			data.NewField(data.TimeSeriesValueFieldName, data.Labels(res.Labels), []*float64{valuePtr}))
-	}
-
-	return frames, nil
-}
-
-func (ls logStats) alertingDataFrames() (data.Frames, error) {
-	frames := make(data.Frames, len(ls.Result))
-	for i, res := range ls.Result {
-		f, err := strconv.ParseFloat(res.Value[1].(string), 64)
-		if err != nil {
-			return nil, fmt.Errorf("metric %v, unable to parse timestamp to float64 from %s: %w", res, res.Value[1], err)
-		}
-
-		frames[i] = data.NewFrame("",
-			data.NewField(data.TimeSeriesValueFieldName, data.Labels(res.Labels), []float64{f})).
-			// to show instant alert response with the table we need to define the type of the frame
-			// and it should be [0, 1] like it set in the Grafana
-			SetMeta(&data.FrameMeta{
-				Type:        data.FrameTypeNumericMulti,
-				TypeVersion: data.FrameTypeVersion{0, 1},
-			})
-	}
-
-	return frames, nil
-}
-
-func (ls logStats) matrixDataFrames() (data.Frames, error) {
-	frames := make(data.Frames, len(ls.Result))
-	for i, res := range ls.Result {
-		timestamps := make([]time.Time, len(res.Values))
-		values := make([]*float64, len(res.Values))
-
-		for j, value := range res.Values {
-			t, err := getTimestamp(value[0])
-			if err != nil {
-				return nil, fmt.Errorf("failed to parse timestamp response for metric %v: %w", res, err)
-			}
-			timestamps[j] = t
-
-			fPtr, err := getFloatPtr(value[1])
-			if err != nil {
-				return nil, fmt.Errorf("failed to parse float value response for metric %v: %w", res, err)
-			}
-			values[j] = fPtr
-		}
-
-		if len(values) < 1 || len(timestamps) < 1 {
-			return nil, fmt.Errorf("log %v contains no values", res)
-		}
-
-		frames[i] = data.NewFrame("",
-			data.NewField(data.TimeSeriesTimeFieldName, nil, timestamps),
-			data.NewField(data.TimeSeriesValueFieldName, data.Labels(res.Labels), values))
-	}
-
-	return frames, nil
-}
-
-func (r *Response) getDataFrames() (data.Frames, error) {
-	var ls logStats
-	if err := json.Unmarshal(r.Data.Result, &ls.Result); err != nil {
-		return nil, fmt.Errorf("unmarshal err %s; \n %#v", err, string(r.Data.Result))
-	}
-
-	switch r.Data.ResultType {
-	case vector:
-		if r.ForAlerting {
-			return ls.alertingDataFrames()
-		}
-		return ls.vectorDataFrames()
-	case matrix:
-		return ls.matrixDataFrames()
-	default:
-		return nil, fmt.Errorf("unknown result type %q", r.Data.ResultType)
-	}
-}
-
-// Hit represents a single hit from the query
-type Hit struct {
-	Fields     map[string]string `json:"fields"`
-	Timestamps []string          `json:"timestamps"`
-	Values     []float64         `json:"values"`
-	Total      int               `json:"total"`
-}
-
-// HitsResponse represents response from the hits query
-type HitsResponse struct {
-	Hits []Hit `json:"hits"`
-}
-
-func (hr *HitsResponse) getDataFrames() (data.Frames, error) {
-	frames := make(data.Frames, len(hr.Hits))
-	for i, hit := range hr.Hits {
-		if len(hit.Timestamps) != len(hit.Values) {
-			return nil, fmt.Errorf("timestamps and values length mismatch: %d != %d", len(hit.Timestamps), len(hit.Values))
-		}
-
-		timeFd := data.NewFieldFromFieldType(data.FieldTypeTime, len(hit.Timestamps))
-		timeFd.Name = gTimeField
-
-		valueFd := data.NewFieldFromFieldType(data.FieldTypeFloat64, len(hit.Values))
-		valueFd.Name = gValueField
-		valueFd.Labels = make(data.Labels)
-
-		for j, ts := range hit.Timestamps {
-			getTime, err := utils.GetTime(ts)
-			if err != nil {
-				return nil, fmt.Errorf("error parse time from _time field: %s", err)
-			}
-			timeFd.Set(j, getTime)
-		}
-
-		for k, v := range hit.Values {
-			valueFd.Set(k, v)
-		}
-
-		for key, value := range hit.Fields {
-			valueFd.Labels[key] = value
-			d, err := labelsToJSON(valueFd.Labels)
-			if err != nil {
-				return nil, fmt.Errorf("error convert labels to json: %s", err)
-			}
-			valueFd.Config = &data.FieldConfig{DisplayNameFromDS: string(d)}
-		}
-
-		frames[i] = data.NewFrame("", timeFd, valueFd)
-	}
-
-	return frames, nil
-}
-
-func getTimestamp(value interface{}) (time.Time, error) {
-	v, ok := value.(float64)
-	if !ok {
-		return time.Time{}, fmt.Errorf("failed to convert timestamp to float64 for value %v", value)
-	}
-
-	seconds := int64(v)                                // get only seconds
-	nanoseconds := int64((v - float64(seconds)) * 1e9) // get only nanoseconds
-	t := time.Unix(seconds, nanoseconds)
-	return t, nil
-}
-
-func getFloatPtr(value interface{}) (*float64, error) {
-	f, ok := value.(string)
-	if !ok {
-		return nil, fmt.Errorf("unable to convert log value to string from %v", value)
-	}
-
-	if f == "" {
-		return nil, nil
-	}
-
-	flVal, err := strconv.ParseFloat(f, 64)
-	if err != nil {
-		return nil, fmt.Errorf("unable to parse log value to float64 from %v: %w", value, err)
-	}
-
-	floatPtr := utils.Ptr(flVal)
-	return floatPtr, nil
 }
