+++ pkg/plugin/response_logs_test.go
@@ -13,8 +13,6 @@ import (
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	"github.com/grafana/grafana-plugin-sdk-go/data"
 	"github.com/valyala/fastjson"
-
-	"github.com/VictoriaMetrics/victorialogs-datasource/pkg/utils"
 )
 
 // newIDField creates an id field pre-populated for the provided rows.
@@ -960,380 +958,6 @@ func Test_parseInstantResponse(t *testing.T) {
 			rsp.Frames = append(rsp.Frames, frame)
 
 			return rsp
-		},
-	}
-	f(o)
-}
-
-func Test_getStatsResponse(t *testing.T) {
-	type opts struct {
-		filename string
-		q        *Query
-		want     func() backend.DataResponse
-	}
-	f := func(opts opts) {
-		t.Helper()
-		file, err := os.ReadFile(opts.filename)
-		if err != nil {
-			t.Fatalf("error reading file: %s", err)
-		}
-
-		r := io.NopCloser(bytes.NewBuffer(file))
-		w := opts.want()
-		resp := parseStatsResponse(r, opts.q)
-
-		if w.Error != nil {
-			if w.Error.Error() != resp.Error.Error() {
-				t.Errorf("parseStreamResponse() = %#v, want %#v", resp, w)
-			}
-			return
-		}
-
-		if len(resp.Frames) != 0 && len(w.Frames) != 0 {
-			got, err := resp.MarshalJSON()
-			if err != nil {
-				t.Fatalf("error marshal response: %s", err)
-			}
-			want, err := w.MarshalJSON()
-			if err != nil {
-				t.Fatalf("error marshal want response: %s", err)
-			}
-			if !bytes.Equal(got, want) {
-				t.Fatalf("\n got value: %s, \n want value: %s", got, want)
-			}
-		}
-
-		if len(resp.Frames) == 0 && len(w.Frames) == 0 {
-			got, err := resp.MarshalJSON()
-			if err != nil {
-				t.Fatalf("error marshal response: %s", err)
-			}
-			want, err := w.MarshalJSON()
-			if err != nil {
-				t.Fatalf("error marshal want response: %s", err)
-			}
-			if !bytes.Equal(got, want) {
-				t.Fatalf("\n got value: %s, \n want value: %s", got, want)
-			}
-		}
-	}
-
-	// empty response
-	o := opts{
-		filename: "test-data/stats_empty",
-		want: func() backend.DataResponse {
-
-			frame := data.NewFrame("", nil)
-
-			rsp := backend.DataResponse{}
-			frame.Meta = &data.FrameMeta{}
-			rsp.Frames = append(rsp.Frames, frame)
-
-			return rsp
-		},
-		q: &Query{},
-	}
-	f(o)
-
-	// incorrect response
-	o = opts{
-		filename: "test-data/stats_incorrect_response",
-		want: func() backend.DataResponse {
-			return newResponseError(fmt.Errorf("failed to prepare data from response: unmarshal err json: cannot unmarshal string into Go value of type []plugin.Result; \n \"\\\"abc\\\"\""), backend.StatusInternal)
-		},
-		q: &Query{},
-	}
-	f(o)
-
-	// correct stats response
-	o = opts{
-		filename: "test-data/stats_response",
-		q: &Query{
-			DataQuery: backend.DataQuery{
-				RefID: "A",
-			},
-			LegendFormat: "legend {{app}}",
-		},
-		want: func() backend.DataResponse {
-			frames := []*data.Frame{
-				data.NewFrame("legend ",
-					data.NewField(data.TimeSeriesTimeFieldName, nil, []time.Time{time.Unix(1730937600, 0)}),
-					data.NewField(data.TimeSeriesValueFieldName, data.Labels{"__name__": "count(*)", "type": "message"}, []*float64{utils.Ptr(float64(13377))}).SetConfig(&data.FieldConfig{DisplayNameFromDS: "legend "}),
-				),
-				data.NewFrame("legend ",
-					data.NewField(data.TimeSeriesTimeFieldName, nil, []time.Time{time.Unix(1730937600, 0)}),
-					data.NewField(data.TimeSeriesValueFieldName, data.Labels{"__name__": "count(*)", "type": ""}, []*float64{utils.Ptr(float64(2078793288))}).SetConfig(&data.FieldConfig{DisplayNameFromDS: "legend "}),
-				),
-			}
-
-			rsp := backend.DataResponse{}
-			rsp.Frames = append(rsp.Frames, frames...)
-			return rsp
-		},
-	}
-	f(o)
-
-	// correct range response
-	o = opts{
-		filename: "test-data/stats_range_response",
-		q: &Query{
-			DataQuery: backend.DataQuery{
-				RefID: "A",
-			},
-			LegendFormat: "legend {{app}}",
-		},
-		want: func() backend.DataResponse {
-			frames := []*data.Frame{
-				data.NewFrame("legend ",
-					data.NewField(data.TimeSeriesTimeFieldName, nil, []time.Time{time.Unix(1704067200, 0), time.Unix(1704088800, 0), time.Unix(1704110400, 0), time.Unix(1704132000, 0)}),
-					data.NewField(data.TimeSeriesValueFieldName, data.Labels{"__name__": "count(*)", "type": ""}, []*float64{utils.Ptr(float64(1311461)), utils.Ptr(float64(1311601)), utils.Ptr(float64(1310266)), utils.Ptr(float64(1310875))}).SetConfig(&data.FieldConfig{DisplayNameFromDS: "legend "}),
-				),
-			}
-
-			rsp := backend.DataResponse{}
-			rsp.Frames = append(rsp.Frames, frames...)
-			return rsp
-		},
-	}
-	f(o)
-
-	// response with milliseconds in timestamps
-	o = opts{
-		filename: "test-data/stats_response_milliseconds",
-		q: &Query{
-			DataQuery: backend.DataQuery{
-				RefID: "A",
-			},
-			LegendFormat: "legend {{app}}",
-			Step:         "10ms",
-		},
-		want: func() backend.DataResponse {
-			frames := []*data.Frame{
-				data.NewFrame("legend ",
-					data.NewField(data.TimeSeriesTimeFieldName, nil, []time.Time{
-						time.Unix(1733187134, 0),
-						time.Unix(1733187134, 449999809),
-					}),
-					data.NewField(data.TimeSeriesValueFieldName, data.Labels{"__name__": "count(*)"}, []*float64{utils.Ptr(float64(58)), utils.Ptr(float64(1))}).SetConfig(&data.FieldConfig{DisplayNameFromDS: "legend "}),
-				),
-			}
-			rsp := backend.DataResponse{}
-			rsp.Frames = append(rsp.Frames, frames...)
-			return rsp
-		},
-	}
-	f(o)
-
-	// correct stats response for alerting
-	o = opts{
-		filename: "test-data/stats_response",
-		q: &Query{
-			DataQuery: backend.DataQuery{
-				RefID: "A",
-			},
-			LegendFormat: "legend {{app}}",
-			ForAlerting:  true,
-		},
-		want: func() backend.DataResponse {
-			frames := []*data.Frame{
-				data.NewFrame("",
-					data.NewField(data.TimeSeriesValueFieldName, data.Labels{"__name__": "count(*)", "type": "message"}, []float64{13377}),
-				).SetMeta(&data.FrameMeta{Type: data.FrameTypeNumericMulti, TypeVersion: data.FrameTypeVersion{0, 1}}),
-				data.NewFrame("",
-					data.NewField(data.TimeSeriesValueFieldName, data.Labels{"__name__": "count(*)", "type": ""}, []float64{2078793288}),
-				).SetMeta(&data.FrameMeta{Type: data.FrameTypeNumericMulti, TypeVersion: data.FrameTypeVersion{0, 1}}),
-			}
-
-			rsp := backend.DataResponse{}
-			rsp.Frames = append(rsp.Frames, frames...)
-			return rsp
-		},
-	}
-	f(o)
-
-	// add interval to config
-	o = opts{
-		filename: "test-data/stats_response",
-		q: &Query{
-			DataQuery: backend.DataQuery{
-				RefID: "A",
-			},
-			LegendFormat: "legend {{app}}",
-			IntervalMs:   1000,
-		},
-		want: func() backend.DataResponse {
-			frames := []*data.Frame{
-				data.NewFrame("legend ",
-					data.NewField(data.TimeSeriesTimeFieldName, nil, []time.Time{time.Unix(1730937600, 0)}).SetConfig(&data.FieldConfig{Interval: 1000}),
-					data.NewField(data.TimeSeriesValueFieldName, data.Labels{"__name__": "count(*)", "type": "message"}, []*float64{utils.Ptr(float64(13377))}).SetConfig(&data.FieldConfig{DisplayNameFromDS: "legend "}),
-				),
-				data.NewFrame("legend ",
-					data.NewField(data.TimeSeriesTimeFieldName, nil, []time.Time{time.Unix(1730937600, 0)}).SetConfig(&data.FieldConfig{Interval: 1000}),
-					data.NewField(data.TimeSeriesValueFieldName, data.Labels{"__name__": "count(*)", "type": ""}, []*float64{utils.Ptr(float64(2078793288))}).SetConfig(&data.FieldConfig{DisplayNameFromDS: "legend "}),
-				),
-			}
-
-			rsp := backend.DataResponse{}
-			rsp.Frames = append(rsp.Frames, frames...)
-			return rsp
-		},
-	}
-	f(o)
-
-	// response with the empty value in the vector response
-	o = opts{
-		filename: "test-data/stats_empty_value_in_the_response",
-		q: &Query{
-			DataQuery: backend.DataQuery{
-				RefID: "A",
-			},
-			LegendFormat: "legend {{app}}",
-		},
-		want: func() backend.DataResponse {
-			frames := []*data.Frame{
-				data.NewFrame("legend ",
-					data.NewField(data.TimeSeriesTimeFieldName, nil, []time.Time{time.Unix(1756992432, 0)}),
-					data.NewField(data.TimeSeriesValueFieldName, data.Labels{"__name__": "p95"}, []*float64{nil}).SetConfig(&data.FieldConfig{DisplayNameFromDS: "legend "}),
-				),
-			}
-
-			rsp := backend.DataResponse{}
-			rsp.Frames = append(rsp.Frames, frames...)
-			return rsp
-		},
-	}
-	f(o)
-
-	// response with the nil value in the vector response
-	o = opts{
-		filename: "test-data/stats_nil_value_in_the_response",
-		q: &Query{
-			DataQuery: backend.DataQuery{
-				RefID: "A",
-			},
-			LegendFormat: "legend {{app}}",
-		},
-		want: func() backend.DataResponse {
-			frames := []*data.Frame{
-				data.NewFrame("legend ",
-					data.NewField(data.TimeSeriesTimeFieldName, nil, []time.Time{time.Unix(1756992432, 0)}),
-					data.NewField(data.TimeSeriesValueFieldName, data.Labels{"__name__": "p95"}, []*float64{nil}).SetConfig(&data.FieldConfig{DisplayNameFromDS: "legend "}),
-				),
-			}
-
-			rsp := backend.DataResponse{}
-			rsp.Frames = append(rsp.Frames, frames...)
-			return rsp
-		},
-	}
-	f(o)
-}
-
-func Test_parseHitsResponse(t *testing.T) {
-	type opts struct {
-		reader io.Reader
-		want   func() backend.DataResponse
-	}
-	f := func(opts opts) {
-		t.Helper()
-		w := opts.want()
-		resp := parseHitsResponse(opts.reader)
-
-		if w.Error != nil {
-			if w.Error.Error() != resp.Error.Error() {
-				t.Errorf("parseStreamResponse() = %#v, want %#v", resp, w)
-			}
-			return
-		}
-
-		if len(resp.Frames) != 0 && len(w.Frames) != 0 {
-			got, err := resp.MarshalJSON()
-			if err != nil {
-				t.Fatalf("error marshal response: %s", err)
-			}
-			want, err := w.MarshalJSON()
-			if err != nil {
-				t.Fatalf("error marshal want response: %s", err)
-			}
-			if !bytes.Equal(got, want) {
-				t.Fatalf("\n got value: %s, \n want value: %s", got, want)
-			}
-		}
-	}
-
-	// empty response
-	o := opts{
-		reader: bytes.NewBufferString(`{ "hits": [] }`),
-		want: func() backend.DataResponse {
-			return backend.DataResponse{}
-		},
-	}
-	f(o)
-
-	// single hit response
-	o = opts{
-		reader: bytes.NewBufferString(`{ "hits": [{ "fields": { "field1": "value1" }, "timestamps": ["2024-01-01T00:00:00Z"], "values": [1.23] }] }`),
-		want: func() backend.DataResponse {
-			timeFd := data.NewFieldFromFieldType(data.FieldTypeTime, 0)
-			timeFd.Name = gTimeField
-			timeFd.Append(time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC))
-
-			valueFd := data.NewFieldFromFieldType(data.FieldTypeFloat64, 0)
-			valueFd.Name = gValueField
-			valueFd.Append(1.23)
-			valueFd.Labels = data.Labels{"field1": "value1"}
-			d, _ := labelsToJSON(valueFd.Labels)
-
-			valueFd.Config = &data.FieldConfig{DisplayNameFromDS: string(d)}
-
-			frame := data.NewFrame("", timeFd, valueFd)
-			return backend.DataResponse{Frames: data.Frames{frame}}
-		},
-	}
-	f(o)
-
-	// multiple hits response
-	o = opts{
-		reader: bytes.NewBufferString(`{ "hits": [{ "fields": { "field1": "value1" }, "timestamps": ["2024-01-01T00:00:00Z"], "values": [1.23] }, { "fields": { "field2": "value2" }, "timestamps": ["2024-01-01T01:00:00Z"], "values": [4.56] }] }`),
-		want: func() backend.DataResponse {
-			timeFd1 := data.NewFieldFromFieldType(data.FieldTypeTime, 0)
-			timeFd1.Name = gTimeField
-			timeFd1.Append(time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC))
-
-			valueFd1 := data.NewFieldFromFieldType(data.FieldTypeFloat64, 0)
-			valueFd1.Name = gValueField
-			valueFd1.Append(1.23)
-			valueFd1.Labels = data.Labels{"field1": "value1"}
-			d, _ := labelsToJSON(valueFd1.Labels)
-
-			valueFd1.Config = &data.FieldConfig{DisplayNameFromDS: string(d)}
-
-			frame1 := data.NewFrame("", timeFd1, valueFd1)
-
-			timeFd2 := data.NewFieldFromFieldType(data.FieldTypeTime, 0)
-			timeFd2.Name = gTimeField
-			timeFd2.Append(time.Date(2024, 1, 1, 1, 0, 0, 0, time.UTC))
-
-			valueFd2 := data.NewFieldFromFieldType(data.FieldTypeFloat64, 0)
-			valueFd2.Name = gValueField
-			valueFd2.Append(4.56)
-			valueFd2.Labels = data.Labels{"field2": "value2"}
-			d, _ = labelsToJSON(valueFd2.Labels)
-
-			valueFd2.Config = &data.FieldConfig{DisplayNameFromDS: string(d)}
-
-			frame2 := data.NewFrame("", timeFd2, valueFd2)
-
-			return backend.DataResponse{Frames: data.Frames{frame1, frame2}}
-		},
-	}
-	f(o)
-
-	// error in response
-	o = opts{
-		reader: bytes.NewBufferString(`{ "hits": [{ "fields": { "field1": "value1" }, "timestamps": ["invalid-time"], "values": [1.23] }] }`),
-		want: func() backend.DataResponse {
-			return newResponseError(fmt.Errorf("failed to prepare data from response: error parse time from _time field: cannot parse invalid-time: cannot parse duration \"invalid-time\""), backend.StatusInternal)
 		},
 	}
 	f(o)
